+++ src/BinaryField.hs
@@ -2,22 +2,24 @@ module BinaryField
   ( BinaryField
   ) where
 
-import Protolude
+import Protolude as P hiding (Semiring)
 
-import Control.Monad.Random (Random(..), getRandom)
+import Control.Monad.Random (Random(..))
+import Data.Euclidean (Euclidean(..), GcdDomain(..))
+import Data.Semiring (Ring(..), Semiring(..))
 import Test.Tasty.QuickCheck (Arbitrary(..), choose)
 import Text.PrettyPrint.Leijen.Text (Pretty(..))
 
-import GaloisField (GaloisField(..))
+import GaloisField (Field(..), GaloisField(..))
 
 -------------------------------------------------------------------------------
--- Binary field type
+-- Data types
 -------------------------------------------------------------------------------
 
 -- | Binary fields @GF(2^q)[X]/\<f(X)\>@ for @q@ positive and
 -- @f(X)@ irreducible monic in @GF(2^q)[X]@ encoded as an integer.
 newtype BinaryField (im :: Nat) = BF Integer
-  deriving (Eq, Generic, NFData, Read, Show)
+  deriving (Eq, Generic, Ord, Show)
 
 -- Binary fields are Galois fields.
 instance KnownNat im => GaloisField (BinaryField im) where
@@ -27,46 +29,27 @@ instance KnownNat im => GaloisField (BinaryField im) where
   {-# INLINE deg #-}
   frob          = flip pow 2
   {-# INLINE frob #-}
-  pow w@(BF y) n
-    | n < 0     = pow (recip w) (-n)
-    | otherwise = BF (pow' 1 y n)
-    where
-      mul = (.) (binMod (natVal w)) . binMul
-      pow' ws zs m
-        | m == 0    = ws
-        | m == 1    = mul ws zs
-        | even m    = pow' ws (mul zs zs) (div m 2)
-        | otherwise = pow' (mul ws zs) (mul zs zs) (div m 2)
-  {-# INLINE pow #-}
-  quad a b c
-    | b == 0    = sr c
-    | otherwise = (* (b / a)) <$> binQuad (a * c / (b * b))
-  {-# INLINE quad #-}
-  rnd           = getRandom
-  {-# INLINE rnd #-}
-  sr            = panic "not implemented."
-  {-# INLINE sr #-}
+
+{-# RULES "BinaryField/pow"
+  forall (k :: KnownNat im => BinaryField im) n . (^) k n = pow k n
+  #-}
 
 -------------------------------------------------------------------------------
--- Binary field instances
+-- Numeric instances
 -------------------------------------------------------------------------------
 
--- Binary fields are arbitrary.
-instance KnownNat im => Arbitrary (BinaryField im) where
-  arbitrary = BF <$> choose (0, order (witness :: BinaryField im) - 1)
-
--- Binary fields are fields.
+-- Binary fields are fractional.
 instance KnownNat im => Fractional (BinaryField im) where
-  recip w@(BF x)      = BF (binInv x (natVal w))
+  recip (BF x)        = BF (binInv x (natVal (witness :: BinaryField im)))
   {-# INLINE recip #-}
   fromRational (x:%y) = fromInteger x / fromInteger y
   {-# INLINABLE fromRational #-}
 
--- Binary fields are rings.
+-- Binary fields are numeric.
 instance KnownNat im => Num (BinaryField im) where
   BF x + BF y = BF (xor x y)
   {-# INLINE (+) #-}
-  BF x * BF y = fromInteger (binMul x y)
+  BF x * BF y = BF (binMul (natVal (witness :: BinaryField im)) x y)
   {-# INLINE (*) #-}
   BF x - BF y = BF (xor x y)
   {-# INLINE (-) #-}
@@ -76,6 +59,53 @@ instance KnownNat im => Num (BinaryField im) where
   {-# INLINABLE fromInteger #-}
   abs         = panic "not implemented."
   signum      = panic "not implemented."
+
+-------------------------------------------------------------------------------
+-- Semiring instances
+-------------------------------------------------------------------------------
+
+-- Binary fields are Euclidean domains.
+instance KnownNat im => Euclidean (BinaryField im) where
+  quotRem = (flip (,) 0 .) . (/)
+  {-# INLINE quotRem #-}
+  degree  = panic "not implemented."
+  {-# INLINE degree #-}
+
+-- Binary fields are fields.
+instance KnownNat im => Field (BinaryField im) where
+  invert = recip
+  {-# INLINE invert #-}
+  minus  = (-)
+  {-# INLINE minus #-}
+
+-- Binary fields are GCD domains.
+instance KnownNat im => GcdDomain (BinaryField im)
+
+-- Binary fields are rings.
+instance KnownNat im => Ring (BinaryField im) where
+  negate = P.negate
+  {-# INLINE negate #-}
+
+-- Binary fields are semirings.
+instance KnownNat im => Semiring (BinaryField im) where
+  zero        = 0
+  {-# INLINE zero #-}
+  plus        = (+)
+  {-# INLINE plus #-}
+  one         = 1
+  {-# INLINE one #-}
+  times       = (*)
+  {-# INLINE times #-}
+  fromNatural = fromIntegral
+  {-# INLINE fromNatural #-}
+
+-------------------------------------------------------------------------------
+-- Other instances
+-------------------------------------------------------------------------------
+
+-- Binary fields are arbitrary.
+instance KnownNat im => Arbitrary (BinaryField im) where
+  arbitrary = BF <$> choose (0, order (witness :: BinaryField im) - 1)
 
 -- Binary fields are pretty.
 instance KnownNat im => Pretty (BinaryField im) where
@@ -88,7 +118,7 @@ instance KnownNat im => Random (BinaryField im) where
   randomR = panic "not implemented."
 
 -------------------------------------------------------------------------------
--- Binary field arithmetic
+-- Binary arithmetic
 -------------------------------------------------------------------------------
 
 -- Binary logarithm.
@@ -99,17 +129,17 @@ binLog = binLog' 2
     binLog' p x
       | x < p     = 0
       | otherwise = case binLog' (p * p) x of
-        l -> let l' = 2 * l in binLog'' (quot x (p ^ l')) l'
+        l -> let l' = 2 * l in binLog'' (P.quot x (p ^ l')) l'
       where
         binLog'' :: Integer -> Int -> Int
         binLog'' y n
           | y < p     = n
-          | otherwise = binLog'' (quot y p) (n + 1)
+          | otherwise = binLog'' (P.quot y p) (n + 1)
 {-# INLINE binLog #-}
 
 -- Binary multiplication.
-binMul :: Integer -> Integer -> Integer
-binMul = binMul' 0
+binMul :: Integer -> Integer -> Integer -> Integer
+binMul = (. binMul' 0) . (.) . binMod
   where
     binMul' :: Integer -> Integer -> Integer -> Integer
     binMul' n x y
@@ -147,20 +177,3 @@ binInv f x = case binInv' 0 1 x f of
       where
         q = max 0 (binLog r - binLog r') :: Int
 {-# INLINE binInv #-}
-
--------------------------------------------------------------------------------
--- Binary field quadratics
--------------------------------------------------------------------------------
-
--- Binary quadratic @y^2+y+x=0@.
-binQuad :: forall im . KnownNat im
-  => BinaryField im -> Maybe (BinaryField im)
-binQuad x
-  | sum xs /= 0 = Nothing
-  | odd m       = Just (sum h)
-  | otherwise   = panic "not implemented."
-  where
-    m  = deg x :: Int
-    xs = take m (iterate (^ (2 :: Int)) x) :: [BinaryField im]
-    h  = zipWith ($) (cycle [identity, const 0]) xs :: [BinaryField im]
-{-# INLINE binQuad #-}