+++ src/ExtensionField.hs
@@ -1,230 +1,192 @@
 module ExtensionField
   ( ExtensionField
+  , PolynomialRing
   , IrreducibleMonic(split)
   , fromField
-  , fromList
-  , t
-  , x
+  , toField
+  , pattern X
+  , pattern X2
+  , pattern X3
+  , pattern Y
   ) where
 
-import Protolude
+import Protolude as P hiding (Semiring, quot, quotRem, rem)
 
-import Control.Monad.Random (Random(..), getRandom)
+import Control.Monad.Random (Random(..))
+import Data.Euclidean (Euclidean(..), GcdDomain(..))
+import Data.Poly.Semiring (VPoly, leading, monomial, scale, toPoly, unPoly, pattern X)
+import Data.Semiring as S (Ring(..), Semiring(..))
+import Data.Vector (fromList)
 import Test.Tasty.QuickCheck (Arbitrary(..), vector)
 import Text.PrettyPrint.Leijen.Text (Pretty(..))
 
-import GaloisField (GaloisField(..))
+import GaloisField (Field(..), GaloisField(..))
 
 -------------------------------------------------------------------------------
--- Extension field type
+-- Data types
 -------------------------------------------------------------------------------
 
 -- | Extension fields @GF(p^q)[X]/\<f(X)\>@ for @p@ prime, @q@ positive, and
 -- @f(X)@ irreducible monic in @GF(p^q)[X]@.
-newtype ExtensionField k im = EF (Polynomial k)
-  deriving (Eq, Generic, NFData, Read, Show)
+newtype ExtensionField k im = EF (VPoly k)
+  deriving (Eq, Generic, Ord, Show)
+
+-- | Polynomial rings.
+type PolynomialRing = VPoly
 
 -- | Irreducible monic splitting polynomial @f(X)@ of extension field.
-class IrreducibleMonic k im where
+class GaloisField k => IrreducibleMonic k im where
   {-# MINIMAL split #-}
   -- | Splitting polynomial @f(X)@.
-  split :: ExtensionField k im -> Polynomial k
-  -- | Splitting polynomial list.
-  plist :: ExtensionField k im -> [k]
-  plist = (\(X xs) -> xs) . split
+  split :: ExtensionField k im -> VPoly k
+  -- | Splitting polynomial degree.
+  deg' :: ExtensionField k im -> Int
+  deg' = pred . fromIntegral . degree . split
 
 -- Extension fields are Galois fields.
-instance (GaloisField k, IrreducibleMonic k im)
-  => GaloisField (ExtensionField k im) where
+instance IrreducibleMonic k im => GaloisField (ExtensionField k im) where
   char          = const (char (witness :: k))
   {-# INLINE char #-}
-  deg w         = deg (witness :: k) * (length (plist w) - 1)
+  deg           = (deg (witness :: k) *) . deg'
   {-# INLINE deg #-}
   frob          = pow <*> char
   {-# INLINE frob #-}
-  pow w@(EF (X y)) n
-    | n < 0     = pow (recip w) (-n)
-    | otherwise = EF (X (pow' [1] y n))
-    where
-      mul = (.) (snd . flip polyQR (plist w)) . polyMul
-      pow' ws zs m
-        | m == 0    = ws
-        | m == 1    = mul ws zs
-        | even m    = pow' ws (mul zs zs) (div m 2)
-        | otherwise = pow' (mul ws zs) (mul zs zs) (div m 2)
-  {-# INLINE pow #-}
-  quad          = panic "not implemented."
-  {-# INLINE quad #-}
-  rnd           = getRandom
-  {-# INLINE rnd #-}
-  sr            = panic "not implemented."
-  {-# INLINE sr #-}
+
+{-# RULES "ExtensionField/pow"
+  forall (k :: IrreducibleMonic k im => ExtensionField k im) n . (^) k n = pow k n
+  #-}
 
 -------------------------------------------------------------------------------
--- Extension field conversions
+-- Numeric instances
 -------------------------------------------------------------------------------
 
--- Polynomial rings.
-newtype Polynomial k = X [k]
-  deriving (Eq, Generic, NFData, Read, Show)
+-- Extension fields are fractional.
+instance IrreducibleMonic k im => Fractional (ExtensionField k im) where
+  recip (EF x)        = EF (polyInv x (split (witness :: ExtensionField k im)))
+  {-# INLINE recip #-}
+  fromRational (x:%y) = fromInteger x / fromInteger y
+  {-# INLINABLE fromRational #-}
 
--- Polynomial rings are rings.
-instance GaloisField k => Num (Polynomial k) where
-  X y + X z     = X (polyAdd y z)
+-- Extension fields are numeric.
+instance IrreducibleMonic k im => Num (ExtensionField k im) where
+  EF x + EF y   = EF (plus x y)
   {-# INLINE (+) #-}
-  X y * X z     = X (polyMul y z)
+  EF x * EF y   = EF (rem (times x y) (split (witness :: ExtensionField k im)))
   {-# INLINE (*) #-}
-  X y - X z     = X (polySub y z)
+  EF x - EF y   = EF (x - y)
   {-# INLINE (-) #-}
-  negate (X y)  = X (map negate y)
+  negate (EF x) = EF (S.negate x)
   {-# INLINE negate #-}
-  fromInteger n = X (let m = fromInteger n in if m == 0 then [] else [m])
+  fromInteger   = EF . fromInteger
   {-# INLINABLE fromInteger #-}
   abs           = panic "not implemented."
   signum        = panic "not implemented."
 
--- | Convert from field element to list representation.
-fromField :: ExtensionField k im -> [k]
-fromField (EF (X y)) = y
-{-# INLINABLE fromField #-}
+-------------------------------------------------------------------------------
+-- Semiring instances
+-------------------------------------------------------------------------------
 
--- | Convert from list representation to field element.
-fromList :: forall k im . (GaloisField k, IrreducibleMonic k im)
-  => [k] -> ExtensionField k im
-fromList = EF . X . snd . flip polyQR (plist w) . dropZero
-  where
-    w = witness :: ExtensionField k im
-{-# INLINABLE fromList #-}
+-- Extension fields are Euclidean domains.
+instance IrreducibleMonic k im => Euclidean (ExtensionField k im) where
+  quotRem = (flip (,) 0 .) . (/)
+  {-# INLINE quotRem #-}
+  degree  = panic "not implemented."
+  {-# INLINE degree #-}
 
--- | Descend tower of indeterminate variables.
-t :: Polynomial k -> Polynomial (ExtensionField k im)
-t = X . return . EF
-{-# INLINE t #-}
+-- Extension fields are fields.
+instance IrreducibleMonic k im => Field (ExtensionField k im) where
+  invert = recip
+  {-# INLINE invert #-}
+  minus  = (-)
+  {-# INLINE minus #-}
 
--- | Current indeterminate variable.
-x :: GaloisField k => Polynomial k
-x = X [0, 1]
-{-# INLINE x #-}
+-- Extension fields are GCD domains.
+instance IrreducibleMonic k im => GcdDomain (ExtensionField k im)
+
+-- Extension fields are rings.
+instance IrreducibleMonic k im => Ring (ExtensionField k im) where
+  negate = P.negate
+  {-# INLINE negate #-}
+
+-- Extension fields are semirings.
+instance IrreducibleMonic k im => Semiring (ExtensionField k im) where
+  zero        = 0
+  {-# INLINE zero #-}
+  plus        = (+)
+  {-# INLINE plus #-}
+  one         = 1
+  {-# INLINE one #-}
+  times       = (*)
+  {-# INLINE times #-}
+  fromNatural = fromIntegral
+  {-# INLINE fromNatural #-}
 
 -------------------------------------------------------------------------------
--- Extension field instances
+-- Other instances
 -------------------------------------------------------------------------------
 
 -- Extension fields are arbitrary.
-instance (Arbitrary k, GaloisField k, IrreducibleMonic k im)
-  => Arbitrary (ExtensionField k im) where
-  arbitrary = fromList <$>
-    vector (length (plist (witness :: ExtensionField k im)) - 1)
-
--- Extension fields are fields.
-instance (GaloisField k, IrreducibleMonic k im)
-  => Fractional (ExtensionField k im) where
-  recip w@(EF (X y))  = EF (X (polyInv y (plist w)))
-  {-# INLINE recip #-}
-  fromRational (y:%z) = fromInteger y / fromInteger z
-  {-# INLINABLE fromRational #-}
-
--- Extension fields are rings.
-instance (GaloisField k, IrreducibleMonic k im)
-  => Num (ExtensionField k im) where
-  EF y + EF z             = EF (y + z)
-  {-# INLINE (+) #-}
-  w@(EF (X y)) * EF (X z) = EF (X (snd (polyQR (polyMul y z) (plist w))))
-  {-# INLINE (*) #-}
-  EF y - EF z             = EF (y - z)
-  {-# INLINE (-) #-}
-  negate (EF y)           = EF (-y)
-  {-# INLINE negate #-}
-  fromInteger             = EF . fromInteger
-  {-# INLINABLE fromInteger #-}
-  abs                     = panic "not implemented."
-  signum                  = panic "not implemented."
+instance IrreducibleMonic k im => Arbitrary (ExtensionField k im) where
+  arbitrary = toField <$> vector (deg' (witness :: ExtensionField k im))
 
 -- Extension fields are pretty.
-instance (GaloisField k, IrreducibleMonic k im)
-  => Pretty (ExtensionField k im) where
-  pretty (EF (X y)) = pretty y
+instance IrreducibleMonic k im => Pretty (ExtensionField k im) where
+  pretty (EF x) = pretty (toList (unPoly x))
 
 -- Extension fields are random.
-instance (GaloisField k, IrreducibleMonic k im)
-  => Random (ExtensionField k im) where
-  random  = first (EF . X . dropZero) . unfold (length (plist w) - 1) []
+instance IrreducibleMonic k im => Random (ExtensionField k im) where
+  random  = first toField . unfold (deg' (witness :: ExtensionField k im)) []
     where
-      w = witness :: ExtensionField k im
-      unfold n ys g
-        | n <= 0    = (ys, g)
+      unfold n xs g
+        | n <= 0    = (xs, g)
         | otherwise = case random g of
-          (y, g') -> unfold (n - 1) (y : ys) g'
+        (x, g') -> unfold (n - 1) (x : xs) g'
   {-# INLINE random #-}
   randomR = panic "not implemented."
 
 -------------------------------------------------------------------------------
--- Extension field arithmetic
+-- Type conversions
 -------------------------------------------------------------------------------
 
--- Polynomial drop zeroes.
-dropZero :: GaloisField k => [k] -> [k]
-dropZero = reverse . dropWhile (== 0) . reverse
-{-# INLINABLE dropZero #-}
+-- | Convert from field element to list representation.
+fromField :: ExtensionField k im -> [k]
+fromField (EF x) = toList (unPoly x)
+{-# INLINABLE fromField #-}
 
--- Polynomial addition.
-polyAdd :: GaloisField k => [k] -> [k] -> [k]
-polyAdd ys     []     = ys
-polyAdd []     zs     = zs
-polyAdd (y:ys) (z:zs) = let w  = y + z
-                            ws = polyAdd ys zs
-                        in if w == 0 && null ws then [] else w : ws
-{-# INLINE polyAdd #-}
+-- | Convert from list representation to field element.
+toField :: forall k im . IrreducibleMonic k im => [k] -> ExtensionField k im
+toField = EF . flip rem (split (witness :: ExtensionField k im)) . toPoly . fromList
+{-# INLINABLE toField #-}
 
--- Polynomial multiplication.
-polyMul :: GaloisField k => [k] -> [k] -> [k]
-polyMul _      [] = []
-polyMul []     _  = []
-polyMul (y:ys) zs = let ws  = map (* y) zs
-                        ws' = polyMul ys zs
-                    in if null ys then ws else polyAdd ws (0 : ws')
-{-# INLINE polyMul #-}
+-- | Pattern for @X^2@.
+pattern X2 :: GaloisField k => VPoly k
+pattern X2 <- _ where X2 = toPoly (fromList [0, 0, 1])
 
--- Polynomial subtraction.
-polySub :: GaloisField k => [k] -> [k] -> [k]
-polySub ys     []     = ys
-polySub []     zs     = map negate zs
-polySub (y:ys) (z:zs) = let w  = y - z
-                            ws = polySub ys zs
-                        in if w == 0 && null ws then [] else w : ws
-{-# INLINE polySub #-}
+-- | Pattern for @X^3@.
+pattern X3 :: GaloisField k => VPoly k
+pattern X3 <- _ where X3 = toPoly (fromList [0, 0, 0, 1])
 
--- Polynomial quotient and remainder.
-polyQR :: forall k . GaloisField k => [k] -> [k] -> ([k], [k])
-polyQR ys zs = polyGCD ([], ys)
-  where
-    z = last zs :: k
-    m = length zs :: Int
-    last :: [k] -> k
-    last []     = 0
-    last [w]    = w
-    last (_:ws) = last ws
-    polyGCD :: ([k], [k]) -> ([k], [k])
-    polyGCD qr@(qs, rs)
-      | n < 0     = qr
-      | otherwise = polyGCD (polyAdd qs ts, polySub rs (polyMul ts zs))
-      where
-        r = last rs :: k
-        n = length rs - m :: Int
-        ts = replicate n 0 ++ [r / z] :: [k]
-{-# INLINE polyQR #-}
+-- | Pattern for descending tower of indeterminate variables.
+pattern Y :: IrreducibleMonic k im => VPoly k -> VPoly (ExtensionField k im)
+pattern Y <- _ where Y = monomial 0 . EF
 
--- Polynomial inverse.
-polyInv :: forall k . GaloisField k => [k] -> [k] -> [k]
-polyInv [y] _ = [recip y]
-polyInv ys zs = case extGCD (zs, ys) of
-  ([w], (ws, _)) -> map (/ w) ws
-  _              -> panic "no multiplicative inverse."
-  where
-    extGCD :: ([k], [k]) -> ([k], ([k], [k]))
-    extGCD (y, []) = (y, ([], [1]))
-    extGCD (y, z)  = (g, (polySub v (polyMul u q), u))
-      where
-        (q, r)      = polyQR y z
-        (g, (u, v)) = extGCD (z, r)
+-------------------------------------------------------------------------------
+-- Polynomial arithmetic
+-------------------------------------------------------------------------------
+
+-- Polynomial inversion algorithm.
+polyInv :: GaloisField k => VPoly k -> VPoly k -> VPoly k
+polyInv xs ps = case first leading (polyGCD xs ps) of
+  (Just (0, x), ys) -> scale 0 (recip x) ys
+  _                 -> panic "no multiplicative inverse."
 {-# INLINE polyInv #-}
+
+-- Polynomial extended greatest common divisor algorithm.
+polyGCD :: forall k . GaloisField k => VPoly k -> VPoly k -> (VPoly k, VPoly k)
+polyGCD x y = polyGCD' 0 1 y x
+  where
+    polyGCD' :: VPoly k -> VPoly k -> VPoly k -> VPoly k -> (VPoly k, VPoly k)
+    polyGCD' s _  r 0  = (r, s)
+    polyGCD' s s' r r' = case quot r r' of
+      q -> polyGCD' s' (s - times q s') r' (r - times q r')
+{-# INLINE polyGCD #-}